requirements.txt
pyserial==3.5
influxdb-client==1.36.1